KinovaGen3.jl

Summary

This package provides the analytical kinematic and dynamic modelling of the Kinova Gen3 industrial robot.

The mechanical properties of the robot such as the link lengths, masses, inertia are obtained from the Kinova Gen3 manual.

In the kinematical model, the reference frames for each link of the robot is selected in accordance with the manual. See page 193 of the manual for the reference frames.

The aim of this package is to achieve an implementation level without making dynamic allocations when a function of the module is called. However, not all functions of the modules manages to prevent dynamic memory allocation.

Kinematic Model

Position-level Kinematics

KinovaGen3.forward_kinematics! — Method

forward_kinematics!(q, x, R)

Calculate the position level forward kinematics for Kinova Gen3 robot.

This function does not perform dynamic memory allocation.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

x::MVector{3, Float64}: the end-effector position of the robot [m]

R::MMatrix{3, 3, Float64}: the end-effector orientation of the robot []

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> x = zeros(MVector{3})
julia> R = zeros(MMatrix{3, 3})
julia> forward_kinematics!(q, x, R)
([0.32874611223407973, -0.393073312842223, 0.44415411959033596], [-0.6597396084411711 0.75 -0.047367172745376274; 0.6123724356957945 0.5000000000000001 -0.6123724356957946; -0.4355957403991577 -0.43301270189221924 -0.7891491309924313])

KinovaGen3.forward_kinematics — Method

forward_kinematics(q)

Calculate the position level forward kinematics for Kinova Gen3 robot.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> forward_kinematics(q)
([0.32874611223407973, -0.393073312842223, 0.44415411959033596], [-0.6597396084411711 0.75 -0.047367172745376274; 0.6123724356957945 0.5000000000000001 -0.6123724356957946; -0.4355957403991577 -0.43301270189221924 -0.7891491309924313])

Velocity-level Kinematics

KinovaGen3.jacobian! — Method

jacobian!(q, J)

Calculate the velocity level forward kinematics for Kinova Gen3 robot.

This function does not perform dynamic memory allocation.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

J::MMatrix{6, 7, Float64}: the geometric Jacobian of the robot []

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> J = zeros(MMatrix{6, 7})
julia> jacobian!(q, J)
6×7 MMatrix{6, 7, Float64, 42} with indices SOneTo(6)×SOneTo(7):
 -0.393073   0.159344  -0.330236  -0.26764   -0.0886181  -0.11046    0.0
 -0.328746   0.0       -0.20503    0.10253   -0.0591954   0.10253    0.0
  0.0       -0.328746   0.190662  -0.345176   0.051568   -0.0729318  0.0
  0.0        0.0       -0.5        0.75      -0.433013    0.75       0.0473672
  0.0        1.0        0.0        0.5        0.866025    0.5        0.612372
 -1.0        0.0       -0.866025  -0.433013   0.25       -0.433013   0.789149

KinovaGen3.jacobian — Method

jacobian(q)

Calculate the velocity level forward kinematics for Kinova Gen3 robot.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> jacobian(q)
6×7 MMatrix{6, 7, Float64, 42} with indices SOneTo(6)×SOneTo(7):
 -0.393073   0.159344  -0.330236  -0.26764   -0.0886181  -0.11046    0.0
 -0.328746   0.0       -0.20503    0.10253   -0.0591954   0.10253    0.0
  0.0       -0.328746   0.190662  -0.345176   0.051568   -0.0729318  0.0
  0.0        0.0       -0.5        0.75      -0.433013    0.75       0.0473672
  0.0        1.0        0.0        0.5        0.866025    0.5        0.612372
 -1.0        0.0       -0.866025  -0.433013   0.25       -0.433013   0.789149

KinovaGen3.inverse_kinematics! — Method

inverse_kinematics!(q, xp, qp, J)

Calculate the velocity level inverse kinematics for Kinova Gen3 robot.

This method reduces the number of allocations

Note that LinearAlgebra.pinv cause allocations during runtime

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

xp::SVector{6, Float64}: the end-effector velocities of the robot [m/s]

qp::MVector{6, Float64}: the joint space velocities of the robot [rad/s]

J::MMatrix{6, 7, Float64}: the Jacobian matrix of the robot []

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> xp = SVector(5.519897654292758, 3.8454568904118693, 5.400775529650293, 0.05767137155933982, 0.09150292411182576, 0.33940753291008463)
julia> qp = zeros(MVector{7})
julia> J = zeros(MMatrix{6, 7})
julia> inverse_kinematics!(q, xp, qp, J)
7-element MVector{7, Float64} with indices SOneTo(7):
 -13.741946912596518
   0.24414409836241704
   5.390106099987782
 -17.3961161502307
   2.6801993763554823
  23.166518880789784
  -8.751150401752152

KinovaGen3.inverse_kinematics — Method

inverse_kinematics(q, xp)

Calculate the velocity level inverse kinematics for Kinova Gen3 robot.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

xp::SVector{6, Float64}: the end-effector velocities of the robot [m/s]

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> xp = SVector(5.519897654292758, 3.8454568904118693, 5.400775529650293, 0.05767137155933982, 0.09150292411182576, 0.33940753291008463)
julia> inverse_kinematics(q, xp)
7-element MVector{7, Float64} with indices SOneTo(7):
 -13.741946912596518
   0.24414409836241704
   5.390106099987782
 -17.3961161502307
   2.6801993763554823
  23.166518880789784
  -8.751150401752152

Dynamics Model

KinovaGen3.mass_matrix! — Method

mass_matrix!(q, M)

Calculate the mass matrix of Kinova Gen3 robot for a given configuration.

This function reduces the number of dynamic memory allocation.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

M::MMatrix{7, 7, Float64}: the mass matrix of the robot []

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> M = zeros(MMatrix{7, 7})
julia> mass_matrix!(q, M)
7×7 MMatrix{7, 7, Float64, 49} with indices SOneTo(7)×SOneTo(7):
  0.494054     -0.172401      0.26987      0.097991      0.030743     0.00400175   -0.000391656
 -0.172401      0.652499     -0.235912     0.0970044    -0.0183927    0.00189615    0.000522066
  0.26987      -0.235912      0.234615     0.000851294   0.0278451    0.00085165   -0.000479402
  0.097991      0.0970044     0.000851294  0.238603      0.00116619   0.0346251     0.00192549
  0.030743     -0.0183927     0.0278451    0.00116619    0.00750583   3.56828e-5    0.000463883
  0.00400175    0.00189615    0.00085165   0.0346251     3.56828e-5   0.0126349     0.000656714
 -0.000391656   0.000522066  -0.000479402  0.00192549    0.000463883  0.000656714   0.00067412

KinovaGen3.mass_matrix — Method

mass_matrix(q)

Calculate the mass matrix of Kinova Gen3 robot for a given configuration.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> mass_matrix(q)
7×7 MMatrix{7, 7, Float64, 49} with indices SOneTo(7)×SOneTo(7):
  0.494054     -0.172401      0.26987      0.097991      0.030743     0.00400175   -0.000391656
 -0.172401      0.652499     -0.235912     0.0970044    -0.0183927    0.00189615    0.000522066
  0.26987      -0.235912      0.234615     0.000851294   0.0278451    0.00085165   -0.000479402
  0.097991      0.0970044     0.000851294  0.238603      0.00116619   0.0346251     0.00192549
  0.030743     -0.0183927     0.0278451    0.00116619    0.00750583   3.56828e-5    0.000463883
  0.00400175    0.00189615    0.00085165   0.0346251     3.56828e-5   0.0126349     0.000656714
 -0.000391656   0.000522066  -0.000479402  0.00192549    0.000463883  0.000656714   0.00067412

KinovaGen3.coriolis! — Method

coriolis!(q, qp, C)

Calculate the Coriolis term, C(q, qp)qp, of Kinova Gen3 robot for a given configuration.

This function reduces the number of dynamic memory allocation.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

qp::SVector{7, Float64}: the joint velocities of the robot [rad/s]

C::MVector{7, Float64}: the mass matrix of the robot []

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> qp = SVector(0.08123992859372045, 0.1594522268081302, 0.14487502589784196, 0.09843558598385844, 0.06467638628492739, 0.05374794957035123, 0.07384203846053861)
julia> C = zeros(MVector{7})
julia> coriolis!(q, qp, C)
7-element MVector{7, Float64} with indices SOneTo(7):
  0.004668503395502678
 -0.021228409144909945
  0.009371757499649963
 -0.000447318871758321
  0.0008351413947819465
  0.001688511069591081
  0.00010500683267616118

KinovaGen3.coriolis — Method

coriolis(q, qp)

Calculate the Coriolis term, C(q, qp)qp, of Kinova Gen3 robot for a given configuration.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

qp::SVector{7, Float64}: the joint velocities of the robot [rad/s]

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> qp = SVector(0.08123992859372045, 0.1594522268081302, 0.14487502589784196, 0.09843558598385844, 0.06467638628492739, 0.05374794957035123, 0.07384203846053861)
julia> coriolis(q, qp)
7-element MVector{7, Float64} with indices SOneTo(7):
  0.004668503395502678
 -0.021228409144909945
  0.009371757499649963
 -0.000447318871758321
  0.0008351413947819465
  0.001688511069591081
  0.00010500683267616118

KinovaGen3.gravity! — Method

gravity!(q, G)

Calculate the gravity term of Kinova Gen3 robot for a given configuration.

This function reduces the number of dynamic memory allocation.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

G::MVector{7, Float64}: the gravity term of the robot []

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> G = zeros(MVector{7})
julia> gravity!(q, G)
7-element MVector{7, Float64} with indices SOneTo(7):
   1.7862417274472353e-16
 -10.576350989174017
   3.0274799819651212
  -5.734529044767631
   0.2514842083025938
  -0.42307721718230645
  -0.024979720551009294

KinovaGen3.gravity — Method

gravity(q)

Calculate the gravity term of Kinova Gen3 robot for a given configuration.

Arguments

q::SVector{7, Float64}: the joint angles of the robot [rad]

Examples

julia> q = SVector(0.0, π / 6, π / 3, π / 2, 0.0, π / 4, 0.0)
julia> gravity(q)
7-element MVector{7, Float64} with indices SOneTo(7):
   1.7862417274472353e-16
 -10.576350989174017
   3.0274799819651212
  -5.734529044767631
   0.2514842083025938
  -0.42307721718230645
  -0.024979720551009294